Converts slider inputs to radio buttons (#132)
Because we'd like to be able to not bias the inputs, the form should have "no value" for each question initially. Sliders don't really lend themselves to that.

This changes the sliders out for radio boxes.

In addition, now that we can have "no value" entries, I made moving on to the next screen disabled unless all the questions have values.

I did change the data format for the questions, so perhaps that is not OK, but it appears to send the same data to the server, so maybe it is.

I also adjusted the temperature page to not jump around so much when entering values, by placing a minHeight constraint on it. And I removed the little thermometer icon.

Fixes #96