Modifies the toggle buttons in the questions so they can be unse… (#170)
This makes the toggle buttons that are part of the question scales be toggleable, so that you can just click on your answer to "unset" the answer and remove it from the checkup.

This makes it possible for radio-button style questions to be optional, even if the user has already selected something. Now they can go back and just unselect the ones they don't want to submit.

As part of solving this, one of the tests I did added a "delete" icon next to each question with an answer, and while that worked well, it took up too much space. I think this way is probably more intuitive anyhow.